+++ animory/src/main/java/com/daggle/animory/domain/account/Account.java
@@ -1,0 +1,31 @@
+package com.daggle.animory.domain.account;
+
+import lombok.*;
+
+import javax.persistence.*;
+import javax.validation.constraints.NotNull;
+
+@Getter
+@Entity
+@Builder
+@AllArgsConstructor
+@Table(name = "account")
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+public class Account {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Integer id;
+
+    @NotNull
+    @Column(unique = true)
+    private String email;
+
+    @NotNull
+    private String password;
+
+    @NotNull
+    @Enumerated(EnumType.STRING)
+    private AccountRole role;
+
+
+}
