+++ animory/src/main/java/com/daggle/animory/domain/account/AccountRole.java
@@ -1,0 +1,7 @@
+package com.daggle.animory.domain.account;
+
+public enum AccountRole {
+
+    USER,
+    SHELTER
+}
