+++ animory/src/main/java/com/daggle/animory/domain/pet/PetService.java
@@ -1,0 +1,10 @@
+package com.daggle.animory.domain.pet;
+
+
+import org.springframework.stereotype.Service;
+
+
+@Service
+public class PetService {
+
+}
