+++ animory/src/main/java/com/daggle/animory/domain/pet/entity/AdoptionStatus.java
@@ -1,0 +1,6 @@
+package com.daggle.animory.domain.pet.entity;
+
+public enum AdoptionStatus {
+    YES,
+    NO
+}
