+++ animory/src/main/java/com/daggle/animory/domain/pet/entity/NeutralizationStatus.java
@@ -1,0 +1,7 @@
+package com.daggle.animory.domain.pet.entity;
+
+public enum NeutralizationStatus {
+    YES,
+    NO,
+    UNKNOWN
+}
