+++ animory/src/main/java/com/daggle/animory/domain/pet/entity/PetType.java
@@ -1,0 +1,8 @@
+package com.daggle.animory.domain.pet.entity;
+
+public enum PetType {
+
+    DOG,
+    CAT,
+    ETC
+}
