+++ animory/src/main/java/com/daggle/animory/domain/pet/entity/Sex.java
@@ -1,0 +1,7 @@
+package com.daggle.animory.domain.pet.entity;
+
+public enum Sex {
+
+    MALE,
+    FEMALE
+}
