+++ animory/src/main/java/com/daggle/animory/domain/pet/fileIO/FileStorageService.java
@@ -1,0 +1,7 @@
+package com.daggle.animory.domain.pet.fileIO;
+
+import org.springframework.web.multipart.MultipartFile;
+
+public interface FileStorageService {
+    String storeFile(MultipartFile file);
+}
