+++ animory/src/main/java/com/daggle/animory/domain/shelter/Province.java
@@ -1,0 +1,21 @@
+package com.daggle.animory.domain.shelter;
+
+public enum Province {
+    서울,
+    부산,
+    대구,
+    인천,
+    광주,
+    대전,
+    울산,
+    세종,
+    경기,
+    강원,
+    충북,
+    충남,
+    전북,
+    전남,
+    경북,
+    경남,
+    제주
+}
