+++ animory/src/main/java/com/daggle/animory/domain/shortform/dto/response/CategoryShortFormPage.java
@@ -1,0 +1,12 @@
+package com.daggle.animory.domain.shortform.dto.response;
+
+import java.util.List;
+
+public record CategoryShortFormPage(
+
+    String categoryTitle,
+    List<ShortFormDto> shortForms,
+    boolean hasNext
+) {
+}
+
