+++ animory/src/main/java/com/daggle/animory/domain/shortform/dto/response/HomeShortFormPage.java
@@ -1,0 +1,10 @@
+package com.daggle.animory.domain.shortform.dto.response;
+
+import java.util.List;
+
+public record HomeShortFormPage (
+    List<ShortFormDto> shortForms,
+    boolean hasNext
+){
+
+}
