--- animory/src/main/java/com/daggle/animory/model/AccountRole.java
@@ -1,7 +1,0 @@
-package com.daggle.animory.model;
-
-public enum AccountRole {
-
-    USER,
-    SHELTER
-}
