--- animory/src/main/java/com/daggle/animory/model/Province.java
@@ -1,21 +1,0 @@
-package com.daggle.animory.model;
-
-public enum Province {
-    서울,
-    부산,
-    대구,
-    인천,
-    광주,
-    대전,
-    울산,
-    세종,
-    경기,
-    강원,
-    충북,
-    충남,
-    전북,
-    전남,
-    경북,
-    경남,
-    제주
-}
