--- animory/src/main/java/com/daggle/animory/model/ShelterAddress.java
@@ -1,24 +1,0 @@
-package com.daggle.animory.model;
-
-import lombok.*;
-
-import javax.persistence.Embeddable;
-import javax.validation.constraints.NotNull;
-
-@Getter
-@Builder
-@Embeddable
-@AllArgsConstructor
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class ShelterAddress {
-
-    @NotNull
-    private Province province;
-
-    private String city;
-
-    private String roadName;
-
-    private String detail;
-
-}
