--- animory/src/main/java/com/daggle/animory/model/pet/AdoptionStatus.java
@@ -1,6 +1,0 @@
-package com.daggle.animory.model.pet;
-
-public enum AdoptionStatus {
-    YES,
-    NO
-}
