--- animory/src/main/java/com/daggle/animory/model/pet/NeutralizationStatus.java
@@ -1,7 +1,0 @@
-package com.daggle.animory.model.pet;
-
-public enum NeutralizationStatus {
-    YES,
-    NO,
-    UNKNOWN
-}
