--- animory/src/main/java/com/daggle/animory/model/pet/PetType.java
@@ -1,8 +1,0 @@
-package com.daggle.animory.model.pet;
-
-public enum PetType {
-
-    DOG,
-    CAT,
-    ETC
-}
