--- animory/src/main/java/com/daggle/animory/model/pet/Sex.java
@@ -1,7 +1,0 @@
-package com.daggle.animory.model.pet;
-
-public enum Sex {
-
-    MALE,
-    FEMALE
-}
