--- animory/src/main/java/com/daggle/animory/service/pet/PetService.java
@@ -1,10 +1,0 @@
-package com.daggle.animory.service.pet;
-
-
-import org.springframework.stereotype.Service;
-
-
-@Service
-public class PetService {
-
-}
