--- animory/src/main/java/com/daggle/animory/service/pet/fileIO/FileStorageService.java
@@ -1,7 +1,0 @@
-package com.daggle.animory.service.pet.fileIO;
-
-import org.springframework.web.multipart.MultipartFile;
-
-public interface FileStorageService {
-    String storeFile(MultipartFile file);
-}
