--- animory/src/main/java/com/daggle/animory/shortform/dto/response/CategoryShortFormPage.java
@@ -1,12 +1,0 @@
-package com.daggle.animory.shortform.dto.response;
-
-import java.util.List;
-
-public record CategoryShortFormPage(
-
-    String categoryTitle,
-    List<ShortFormDto> shortForms,
-    boolean hasNext
-) {
-}
-
