--- animory/src/main/java/com/daggle/animory/shortform/dto/response/HomeShortFormPage.java
@@ -1,10 +1,0 @@
-package com.daggle.animory.shortform.dto.response;
-
-import java.util.List;
-
-public record HomeShortFormPage (
-    List<ShortFormDto> shortForms,
-    boolean hasNext
-){
-
-}
